BUGFIX Network speed test

+++ FedAdapt/RL_training/RL_runscheduler.py
@@ -39,42 +39,46 @@ def scope1():
                     config.B = b
                     for d in data_lenght_range:
                         config.N = d
-                        for l in learning_rate_range:
-                            config.LR = l
-                            for m in max_update_epochs_range:
-                                config.max_update_epochs = m
-                                for t in tolerance_range:
-                                    config.tolerance_counts = t
 
-                                    try:
-                                        time_server_start = time.perf_counter()
-                                        print("##########################\nRUN METRICS: \n  E: "+str(e)+" \n  I: "+str(i)+ "\n  B: "+str(b)+" \n  D: "+str(d)+" \n  L: "+str(l)+ "\n  M: "+str(m)+" \n  T: "+str(t)+"\n##########################")   
-                                        #######################             
-                                        run_identifier = "E"+str(e)+"_I"+str(i)+"_B"+str(b)+"_D"+str(d)+"_L"+str(l)+ "_M"+str(m)+"_T"+str(t)
+                        ##MAX ITERATION NUMBER = DATASET_SIZE (d) / ( NUMBER OF DEVICES (K) * BATCH_SIZE (B) ) i.e 5000 / (5 * 100) = 10 Due to implementation issues with trainloader
+                        max_iter_number = d / ( 5 * b )
+                        if i <= max_iter_number:
+                            for l in learning_rate_range:
+                                config.LR = l
+                                for m in max_update_epochs_range:
+                                    config.max_update_epochs = m
+                                    for t in tolerance_range:
+                                        config.tolerance_counts = t
+
                                         try:
-                                            os.remove("PPO.pth") ##Remove old trained model, so it creates a new one,and trains without being influenced by previous trains
-                                        except Exception as exception_file_already_removed:
-                                            pass #file already removed
+                                            time_server_start = time.perf_counter()
+                                            print("##########################\nRUN METRICS: \n  E: "+str(e)+" \n  I: "+str(i)+ "\n  B: "+str(b)+" \n  D: "+str(d)+" \n  L: "+str(l)+ "\n  M: "+str(m)+" \n  T: "+str(t)+"\n##########################")   
+                                            #######################             
+                                            run_identifier = "E"+str(e)+"_I"+str(i)+"_B"+str(b)+"_D"+str(d)+"_L"+str(l)+ "_M"+str(m)+"_T"+str(t)
+                                            try:
+                                                os.remove("PPO.pth") ##Remove old trained model, so it creates a new one,and trains without being influenced by previous trains
+                                            except Exception as exception_file_already_removed:
+                                                pass #file already removed
 
-                                        ####################
-                                        start_run(run_identifier,device_type)
+                                            ####################
+                                            start_run(run_identifier,device_type)
+                                            
+                                        except Exception as exception:
+                                            print("EXCEPTION OCCURED DURING RUN: E"+str(e)+"_I"+str(i)+"_B"+str(b)+"_D"+str(d)+"_L"+str(l)+ "_M"+str(m)+"_T"+str(t))
+                                            print(exception)
+                                            ###TODO: Try with clients; check if server start/ipaddress assignation is a problem
+
                                         
-                                    except Exception as exception:
-                                        print("EXCEPTION OCCURED DURING RUN: E"+str(e)+"_I"+str(i)+"_B"+str(b)+"_D"+str(d)+"_L"+str(l)+ "_M"+str(m)+"_T"+str(t))
-                                        print(exception)
-                                        ###TODO: Try with clients; check if server start/ipaddress assignation is a problem
-
-                                    
-                                    #print("CPU_PERCENT_RUN: "+ str(psutil.cpu_percent()))
-                                    time_server_finish = time.perf_counter()
-                                    print("FINISHED RUN: " + run_identifier)
-                                    print("RUN TIME: "+ str(time_server_finish-time_server_start))
-                                    if device_type == "server":
-                                        print("Waiting 5s for address deallocation...")
-                                        time.sleep(5) #waiting for de-allocation of server address
-                                    elif device_type == "client":
-                                        print("Waiting 8s for address deallocation...")
-                                        time.sleep(8) #waiting for de-allocation of server address
+                                        #print("CPU_PERCENT_RUN: "+ str(psutil.cpu_percent()))
+                                        time_server_finish = time.perf_counter()
+                                        print("FINISHED RUN: " + run_identifier)
+                                        print("RUN TIME: "+ str(time_server_finish-time_server_start))
+                                        if device_type == "server":
+                                            print("Waiting 5s for address deallocation...")
+                                            time.sleep(5) #waiting for de-allocation of server address
+                                        elif device_type == "client":
+                                            print("Waiting 8s for address deallocation...")
+                                            time.sleep(8) #waiting for de-allocation of server address
                                 
 
 def scope2():
